refactor: move `eval_hook` from `nu-cli` to `nu-command` to avoid a circular dependency
