+++ src/test_bins.rs
@@ -1,7 +1,7 @@
 use std::io::{self, BufRead, Read, Write};
 
-use nu_cli::{eval_env_change_hook, eval_hook};
 use nu_command::create_default_context;
+use nu_command::hook::{eval_env_change_hook, eval_hook};
 use nu_engine::eval_block;
 use nu_parser::parse;
 use nu_protocol::engine::{EngineState, Stack, StateWorkingSet};
